backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.20
anthropic==0.40.0
python-dotenv==1.0.1
pydantic==2.10.3
